Documentation improvements and fix Azure (#157)
* Fix config javadoc

* Improve docs, add examples for static content and ordering

* Fix fields requiring a signed 64-bit integer

or double-precision float type in the api docs

* Headers are in lowercase, and response needs to be encoded

* Case-insensitive map and remove jsonassert

* Roll back to v1.0.0-SNAPSHOT

+++ chatbots-telegram-api/src/main/java/io/micronaut/chatbots/telegram/api/Chat.java
@@ -35,7 +35,7 @@ public class Chat {
      */
     @NonNull
     @NotNull
-    private Integer id;
+    private Long id;
 
     /**
      * Type of chat, can be either “private”, “group”, “supergroup” or “channel”.
@@ -192,7 +192,7 @@ public class Chat {
      * @return Unique identifier for this chat.
      */
     @NonNull
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -200,7 +200,7 @@ public class Chat {
      *
      * @param id Unique identifier for this chat.
      */
-    public void setId(@NonNull Integer id) {
+    public void setId(@NonNull Long id) {
         this.id = id;
     }
